add dummy button so that the ButtonGroup class works compatible with gtk+ behavior.
when selecting a radio button in a group:
in gtk+:
1. click a radio button in a group.
2. all buttons once unselected,
3. toggled event signaled on previously selected button.
4. a button clicked is selected.
5. toggle event signaled on newly selected button.

in swing, which a group never becomes with all buttons unselected,
except it's in initial state.
1. click a radio button in a group.
2. state change event signaled on the newly selected radio button.

This difference makes it hard to implement processing of radio button events.
Adding dummy button, which is selected when it is desired no actual buttons
are selected, resolves this problem.